New special event + fix weird custom scale
