+++ APTT_olympic/data/dictionary.csv
@@ -1,0 +1,202 @@
+Country,Code,Population,GDP per Capita
+Afghanistan,AFG,32526562,594.323081219966
+Albania,ALB,2889167,3945.21758150914
+Algeria,ALG,39666519,4206.03123244958
+American Samoa*,ASA,55538,
+Andorra,AND,70473,
+Angola,ANG,25021974,4101.47215182964
+Antigua and Barbuda,ANT,91818,13714.7319616988
+Argentina,ARG,43416755,13431.8783398577
+Armenia,ARM,3017712,3489.12768956995
+Aruba*,ARU,103889,
+Australia,AUS,23781169,56310.9629933721
+Austria,AUT,8611088,43774.985173612
+Azerbaijan,AZE,9651349,5496.34464026248
+Bahamas,BAH,388019,22817.2308572518
+Bahrain,BRN,1377237,22600.2140981035
+Bangladesh,BAN,160995642,1211.70153057661
+Barbados,BAR,284215,15429.3404640853
+Belarus,BLR,9513000,5740.45649479562
+Belgium,BEL,11285721,40324.0277657215
+Belize,BIZ,359287,4878.72126724745
+Bermuda*,BER,65235,
+Benin,BEN,10879829,762.051205441965
+Bhutan,BHU,774830,2655.99889161858
+Bolivia,BOL,10724705,3076.79181060881
+Bosnia and Herzegovina,BIH,3810416,4249.3303131949
+Botswana,BOT,2262485,6360.13822018837
+Brazil,BRA,207847528,8538.5899749574
+British Virgin Islands,IVB,30117,
+Brunei,BRU,423188,30554.7296658073
+Bulgaria,BUL,7177991,6993.47735975728
+Burkina Faso,BUR,18105570,589.774414136617
+Burundi,BDI,11178921,277.068309170914
+Cambodia,CAM,15577899,1158.6899035244
+Cameroon,CMR,23344179,1217.26067048427
+Canada,CAN,35851774,43248.529909341
+Cape Verde,CPV,520502,3080.17881445563
+Cayman Islands*,CAY,59967,
+Central African Republic,CAF,4900274,323.201674024141
+Chad,CHA,14037472,775.695090525313
+Chile,CHI,17948141,13416.2300390185
+China,CHN,1371220000,8027.68381013907
+Colombia,COL,48228704,6056.14771720405
+Comoros,COM,788474,717.448849960912
+Congo,CGO,4620330,1851.19991555492
+"Congo, Dem Rep",COD,77266814,456.052740548027
+Cook Islands,COK,,
+Costa Rica,CRC,4807850,11260.0921598775
+Cote d'Ivoire,CIV,22701556,1398.98995768413
+Croatia,CRO,4224404,11535.8293558997
+Cuba,CUB,11389562,
+Cyprus,CYP,1165300,23242.8400685313
+Czech Republic,CZE,10551219,17548.3382129701
+Denmark,DEN,5676002,51989.2934712354
+Djibouti,DJI,887861,1945.12429310444
+Dominica,DMA,72680,7116.38639189547
+Dominican Republic,DOM,10528391,6468.4734915433
+East Timor (Timor-Leste),TLS,1245015,1157.99295590816
+Ecuador,ECU,16144363,6205.06414529951
+Egypt,EGY,91508084,3614.74676616271
+El Salvador,ESA,6126583,4219.35032953932
+Equatorial Guinea,GEQ,845060,14439.5944478516
+Eritrea,ERI,,
+Estonia,EST,1311998,17118.5042003236
+Ethiopia,ETH,99390750,619.169406475891
+Fiji,FIJ,892145,4960.51995438419
+Finland,FIN,5482013,42311.0362306446
+France,FRA,66808385,36205.5681017036
+Gabon,GAB,1725292,8266.4456050624
+Gambia,GAM,1990924,471.537195472347
+Georgia,GEO,3679000,3795.97330845042
+Germany,GER,81413145,41313.3139945434
+Ghana,GHA,27409893,1369.70112227585
+Greece,GRE,10823732,18002.2305776688
+Grenada,GRN,106825,9212.02035173484
+Guam,GUM,169885,
+Guatemala,GUA,16342897,3903.47885604613
+Guinea,GUI,12608590,531.320595188714
+Guinea-Bissau,GBS,1844325,572.988428543071
+Guyana,GUY,767085,4127.35101806198
+Haiti,HAI,10711067,818.343297634549
+Honduras,HON,8075060,2528.89354988523
+Hong Kong*,HKG,7305700,42327.8399570345
+Hungary,HUN,9844686,12363.5434596539
+Iceland,ISL,330823,50173.3399156473
+India,IND,1311050527,1598.25903421916
+Indonesia,INA,257563815,3346.48703949478
+Iran,IRI,79109272,
+Iraq,IRQ,36423395,4943.76038832056
+Ireland,IRL,4640703,61133.672427322
+Israel,ISR,8380400,35728.0934951515
+Italy,ITA,60802085,29957.8043154372
+Jamaica,JAM,2725941,5232.02458271187
+Japan,JPN,126958472,32477.2151449234
+Jordan,JOR,7594547,4940.04583574111
+Kazakhstan,KAZ,17544126,10509.9810699442
+Kenya,KEN,46050302,1376.71282894881
+"Korea, North",PRK,25155317,
+"Korea, South",KOR,50617045,27221.5240509661
+Kuwait,KUW,3892115,29300.5755750333
+Kyrgyzstan,KGZ,5957000,1103.2153515202
+Laos,LAO,6802023,1818.44137293236
+Latvia,LAT,1978440,13648.5475564772
+Lebanon,LIB,5850743,8047.64508557496
+Lesotho,LES,2135022,1066.98562625017
+Liberia,LBR,4503438,455.873934536237
+Libya,LBA,6278438,
+Liechtenstein,LIE,37531,
+Lithuania,LTU,2910199,14147.0493792403
+Luxembourg,LUX,569676,101449.968168042
+Macedonia,MKD,2078453,4852.6578476699
+Madagascar,MAD,24235390,401.836006029745
+Malawi,MAW,17215232,371.985747810344
+Malaysia,MAS,30331007,9768.32686011881
+Maldives,MDV,409163,8395.78519750383
+Mali,MLI,17599694,724.256283193304
+Malta,MLT,431333,22596.1817742659
+Mauritania,MTN,4067564,
+Mauritius,MRI,1262605,9252.11072429051
+Mexico,MEX,127017224,9005.02426497766
+Micronesia,FSM,104460,3015.23166762397
+Moldova,MDA,3554150,1848.06180430428
+Monaco,MON,37731,
+Mongolia,MGL,2959134,3967.82938559498
+Morocco,MAR,34377511,2878.20134215919
+Mozambique,MOZ,27977863,529.242556068943
+Burma,MYA,53897154,1161.48815791087
+Namibia,NAM,2458830,4673.56724769495
+Nauru,NRU,10222,9827.54428148233
+Nepal,NEP,28513700,743.322965726435
+Netherlands,NED,16936520,44299.768085383
+Netherlands Antilles*,AHO,,
+New Zealand,NZL,4595700,37807.9672760442
+Nicaragua,NCA,6082032,2086.89500277099
+Niger,NIG,19899120,358.958152040005
+Nigeria,NGR,182201962,2640.2907388575
+Norway,NOR,5195921,74400.3697770928
+Oman,OMA,4490541,15550.6762514328
+Pakistan,PAK,188924874,1434.69666504969
+Palau,PLW,21291,13498.661406228
+"Palestine, Occupied Territories",PLE,,
+Panama,PAN,3929141,13268.1137549335
+Papua New Guinea,PNG,7619321,
+Paraguay,PAR,6639123,4080.95144785437
+Peru,PER,31376670,6027.12585529551
+Philippines,PHI,100699395,2904.20208191528
+Poland,POL,37999494,12554.5475536313
+Portugal,POR,10348648,19222.1500764153
+Puerto Rico*,PUR,3474182,
+Qatar,QAT,2235355,73653.3944346574
+Romania,ROM,19832389,8972.92251840971
+Russia,RUS,144096812,9092.58053606884
+Rwanda,RWA,11609666,697.34822808289
+Saint Kitts and Nevis,SKN,55572,15771.9455041308
+Saint Lucia,LCA,184999,7735.91048440102
+Saint Vincent and the Grenadines,VIN,109462,6739.17483286945
+Samoa,SAM,193228,3938.54884570316
+San Marino,SMR,31781,
+Sao Tome and Principe,STP,190344,1669.06326801326
+Saudi Arabia,KSA,31540372,20481.7453220484
+Senegal,SEN,15129273,899.579879484977
+Serbia,SCG,7098247,5235.14220696525
+Seychelles,SEY,92900,15476.01944443
+Sierra Leone,SLE,6453184,653.131815940648
+Singapore,SIN,5535002,52888.7446717529
+Slovakia,SVK,5424050,16088.2775872723
+Slovenia,SLO,2063768,20726.5398863708
+Solomon Islands,SOL,583591,1934.85629287353
+Somalia,SOM,10787104,549.266976567576
+South Africa,RSA,54956920,5723.97335690212
+Spain,ESP,46418269,25831.5823052954
+Sri Lanka,SRI,20966000,3926.17439589454
+Sudan,SUD,40234882,2414.72360102858
+Suriname,SUR,542975,9485.31924429496
+Swaziland,SWZ,1286970,3200.14301756487
+Sweden,SWE,9798871,50579.6736486777
+Switzerland,SUI,8286976,80945.0792194742
+Syria,SYR,18502413,
+Taiwan,TPE,,
+Tajikistan,TJK,8481855,925.91188767081
+Tanzania,TAN,53470420,878.97510552608
+Thailand,THA,67959359,5814.76976382355
+Togo,TOG,7304578,559.635876661945
+Tonga,TGA,106170,4098.54392925825
+Trinidad and Tobago,TRI,1360088,17321.884675056
+Tunisia,TUN,11107800,3872.51208364171
+Turkey,TUR,78665830,9125.68758972936
+Turkmenistan,TKM,5373502,6672.47754417351
+Uganda,UGA,39032383,705.292569535068
+Ukraine,UKR,45198200,2114.95471628444
+United Arab Emirates,UAE,9156963,40438.7629344394
+United Kingdom,GBR,65138232,43875.9696143686
+United States,USA,321418820,56115.7184261955
+Uruguay,URU,3431555,15573.9009191818
+Uzbekistan,UZB,31299500,2132.07036847857
+Vanuatu,VAN,264652,2805.31464353567
+Venezuela,VEN,31108083,
+Vietnam,VIE,91703800,2111.13802366815
+Virgin Islands*,ISV,103574,
+Yemen,YEM,26832215,1406.2916511457
+Zambia,ZAM,16211767,1304.87901447726
+Zimbabwe,ZIM,15602751,924.143819253412
