+++ APTT_olympic/data/stripped.csv
@@ -1,0 +1,152 @@
+Year,Country,Medal,Event
+1968,SWE,Bronze,Modern Pentathlon Team
+1972,MNG,Silver,Judo men's 63kg
+1972,NLD,Bronze,Cycling Men's team time trial 
+1972,ESP,Bronze,Cycling Men's individual road race
+1972,USA,Gold,Swimming Men's 400 m freestyle
+1976,RUS,Bronze,Cross-Country Skiing Women's 5 km
+1976,BGR,Gold,Weightlifting Men's 110 kg
+1976,BGR,Silver,Weightlifting Men's 82.5 kg
+1976,POL,Gold,Weightlifting Men's 67.5 kg
+1984,FIN,Silver,Athletics Men's 10 000 m
+1984,SWE,Silver,Wrestling Men's Greco-Roman +100 kg
+1988,BGR,Gold,Weightlifting Men's 56 kg
+1988,BGR,Gold,Weightlifting Men's 67.5 kg
+1988,CAN,Gold,Athletics Men's 100 m
+1988,HUN,Silver,Weightlifting Men's 100 kg
+1988,GBR,Bronze,Judo Men's 71 kg
+2000,ARM,Bronze,Weightlifting Men's +105 kg
+2000,BGR,Gold,Weightlifting Women's 48 kg
+2000,BGR,Silver,Weightlifting Men's 56 kg
+2000,BGR,Bronze,Weightlifting Men's 62 kg
+2000,CHN,Bronze,Gymnastics Women's artistic team all-around
+2000,GER,Gold,Wrestling Men's freestyle 76 kg
+2000,ROU,Gold,Gymnastics Women's artistic individual all-around
+2000,USA,Gold,Athletics Women's 100 m
+2000,USA,Gold,Athletics Women's 200 m
+2000,USA,Bronze,Athletics Women's long jump
+2000,USA,Gold,Athletics Men's 4x400 m relay
+2000,USA,Bronze,Cycling Men's road time trial
+2002,GBR,Bronze,Alpine Skiing Men's slalom
+2002,RUS,Gold,Cross-Country Skiing Women's 5 km + 5 km combined pursuit
+2002,RUS,Silver,Cross-Country Skiing Women's 10 km classical
+2002,RUS,Gold,Cross-Country Skiing Women's 30 km classical
+2002,RUS,Silver,Cross-Country Skiing Women's 15 km freestyle mass start
+2002,RUS,Silver,Cross-Country Skiing Women's 5 km + 5 km combined pursuit
+2002,ESP,Gold,Cross-Country Skiing Men's 50 km classical
+2002,ESP,Gold,Cross-Country Skiing Men's 30 km freestyle
+2002,ESP,Gold,Cross-Country Skiing Men's 10 km + 10 km combined pursuit
+2004,BLR,Silver,Athletics Men's hammer throw
+2004,BLR,Bronze,Athletics Women's discus throw
+2004,GER,Gold,Equestrian Team show jumping
+2004,GRC,Bronze,Weightlifting Men's 62 kg
+2004,HUN,Gold,Athletics Men's hammer throw
+2004,HUN,Gold,Athletics Men's discus throw
+2004,HUN,Silver,Weightlifting Men's 105 kg
+2004,IRE,Gold,Equestrian Individual show jumping
+2004,RUS,Gold,Athletics Women's shot put
+2004,RUS,Bronze,Athletics Women's shot put
+2004,RUS,Bronze,Weightlifting Men's 77 kg
+2004,UKR,Gold,Athletics Men's shot put
+2004,UKR,Bronze,Rowing Women's quadruple sculls
+2004,USA,Gold,Cycling Men's road time trial
+2006,RUS,Silver,Biathlon Women's individual
+2008,ARM,Bronze,Weightlifting Men's 69 kg
+2008,AZE,Silver,Wrestling Men's Greco-Roman 60 kg
+2008,BHR,Gold,Athletics Men's 1500 m
+2008,BLR,Gold,Athletics Women's hammer throw
+2008,BLR,Silver,Athletics Women's shot put
+2008,BLR,Silver,Weightlifting Men's 85 kg
+2008,BLR,Bronze,Athletics Men's shot put
+2008,BLR,Bronze,Weightlifting Women's 53 kg
+2008,BLR,Bronze,Athletics Women's shot put
+2008,CHN,Gold,Weightlifting Women's 69 kg
+2008,CHN,Gold,Weightlifting Women's 75 kg
+2008,CHN,Gold,Weightlifting Women's 48 kg
+2008,CUB,Silver,Athletics Women's discus throw
+2008,GRE,Bronze,Athletics Women's triple jump
+2008,ITA,Silver,Cycling Men's road race
+2008,JAM,Gold,Athletics Men's 4X100 m relay
+2008,KAZ,Gold,Weightlifting Men's 94 kg
+2008,KAZ,Silver,Weightlifting Women's 63 kg
+2008,KAZ,Silver,Wrestling Men's freestyle 96 kg
+2008,KAZ,Bronze,Weightlifting Women's +75 kg
+2008,KAZ,Bronze,Wrestling Men's Greco-Roman 96 kg
+2008,PKR,Silver,Shooting Men's 50 m air pistol
+2008,PKR,Bronze,Shooting Men's 10 m air pistol
+2008,NOR,Bronze,Equestrian team show jumping
+2008,RUS,Gold,Athletics Women's 4X100 m relay
+2008,RUS,Silver,Athletics Women's javelin throw
+2008,RUS,Silver,Wrestling Men's Greco-Roman 120 kg
+2008,RUS,Silver,Athletics Women's triple jump
+2008,RUS,Silver,Athletics Women's long jump
+2008,RUS,Silver,Athletics Women's 4X400 m relay
+2008,RUS,Silver,Weightlifting Women's 58 kg
+2008,RUS,Bronze,Weightlifting Men's 94 kg
+2008,RUS,Bronze,Athletics Women's high jump
+2008,RUS,Bronze,Weightlifting Women's 75 kg
+2008,RUS,Bronze,Weightlifting Men's 105 kg
+2008,RUS,Bronze,Athletics Women's heptathlon
+2008,RUS,Bronze,Athletics Men's 4x400 m relay
+2008,RUS,Bronze,Athletics Women's 3000 m steeplechase
+2008,SWE,Bronze,Wrestling Men's Greco-Roman 84 kg
+2008,TUR,Silver,Athletics Women's 5000 metres
+2008,TUR,Silver,Athletics Women's 10000 metres
+2008,TUR,Silver,Weightlifting Women's 48 kg
+2008,UKR,Silver,Athletics Women's heptathlon
+2008,UKR,Silver,Wrestling Men's freestyle 60 kg
+2008,UKR,Silver,Weightlifting Women's +75 kg
+2008,UKR,Bronze,Weightlifting Women's 69 kg
+2008,UKR,Bronze,Modern pentathlon Women's
+2008,UKR,Bronze,Athletics Men's pole vault
+2008,UZB,Gold,Wrestling Men's freestyle 120 kg
+2008,UZB,Silver,Wrestling Men's freestyle 74 kg
+2012,ARM,Bronze,Weightlifting Women's +75 kg
+2012,AZE,Bronze,Weightlifting Men's 56 kg
+2012,BEL,Gold,Athletics Women's shot put
+2012,BEL,Bronze,Weightlifting Women's 75 kg
+2012,MBEL,Bronze,Weightlifting Women's 69 kg
+2012,GEO,Silver,Wrestling Men's freestyle 120 kg
+2012,KAZ,Gold,Weightlifting Women's 53 kg
+2012,KAZ,Gold,Weightlifting Men's 94 kg
+2012,KAZ,Gold,Weightlifting Women's 63 kg
+2012,KAZ,Gold,Weightlifting Women's 75 kg
+2012,LTU,Silver,Canoeing Men's C-1 200 m
+2012,MDA,Bronze,Weightlifting Men's 94 kg
+2012,MDA,Bronze,Weightlifting Women's 53 kg
+2012,ROU,Bronze,Weightlifting Men's 69 kg
+2012,ROU,Silver,Weightlifting Women's 69 kg
+2012,RUS,Gold,Athletics Men's 50 km walk
+2012,RUS,Gold,Athletics Women's 20 km walk
+2012,RUS,Gold,Athletics Men's high jump
+2012,RUS,Gold,Athletics Women's hammer throw
+2012,RUS,Gold,Athletics Women's 800 m
+2012,RUS,Gold,Athletics Women's 3000 m steeplechase
+2012,RUS,Silver,Weightlifting Men's 85 kg
+2012,RUS,Silver,Weightlifting Men's 94 kg
+2012,RUS,Silver,Athletics Women's 20 km walk
+2012,RUS,Silver,Athletics Women's shot put
+2012,RUS,Silver,Athletics Women's discus throw 
+2012,RUS,Silver,Athletics Women's 4x400 m relay
+2012,RUS,Silver,Weightlifting Women's 63 kg
+2012,RUS,Silver,Weightlifting Women's 75 kg
+2012,RUS,Bronze,Athletics Women's heptathlon
+2012,RUS,Bronze,Athletics Women's high jump
+2012,TUR,Gold,Athletics Women's 1500 m
+2012,TUR,Silver,Athletics Women's 1500 m
+2012,USA,Silver,Athletics Men's 4x100 m relay
+2012,UKR,Gold,Weightlifting Men's 105 kg
+2012,UKR,Silver,Athletics Men's javelin throw
+2012,UKR,Bronze,Weightlifting Women's 58 kg
+2012,UZB,Gold,Wrestling Men's freestyle 120 kg
+2012,UZB,Bronze,Wrestling Men's freestyle 74 kg
+2014,RUS,Gold,Bobsleigh Two-man
+2014,RUS,Gold,Bobsleigh Four-man
+2014,RUS,Gold,Biathlon Men's relay
+2014,RUS,Silver,Biathlon Women's relay
+2016,KAZ,Gold,Weightlifting Men's 77 kg
+2016,KGZ,Bronze,Weightlifting Men's 69 kg
+2016,MDA,Bronze,Canoeing Men's C-1 1000 m
+2016,ROU,Bronze,Weightlifting Men's 85 kg
+2016,RUS,Silver,Boxing Men's flyweight
+2020,GBR,Silver,Athletics Men's 4x100 m relay
