+++ delta.py
@@ -4,7 +4,7 @@ from dash import html
 from energies import energies
 from population import population
 from deces import deces
-from APTT_sujet import olympics
+from APTT_olympic import olympics
 
 # external_stylesheets = ['https://codepen.io/chriddyp/pen/bWLwgP.css']
